ws-server/requirements.txt
git+https://github.com/dpallot/simple-websocket-server.git#egg=SimpleWebSocketServer
elasticsearch-dsl>=5.0.0,<6.0.0
